backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
supabase==2.13.0
sentence-transformers==3.4.1
google-generativeai==0.8.4
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.6
numpy<2.0
scipy>=1.10.0
torch>=2.0.0
einops>=0.8.0
huggingface_hub>=0.29.0
transformers>=4.41.0
# PDF parsing dependencies
pdfplumber==0.10.4
pytesseract==0.3.10
pdf2image==1.17.0
pandas==2.2.2
loguru==0.7.2
Pillow==10.3.0 